Dank Memes
Skipped ahead and downloaded a section from the tutorial. Has a save load feature, only step now is to add a setting to toggle ui visibility.  figure out how to generate navmeshes at runtime, and then port it to multiplayer as well as have it automatically load maps we want